# backend/openapi-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AIInterviewCoach.API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/interviews": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Interviews"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{id}/problems": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProblemToInterviewRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProblemToInterviewRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddProblemToInterviewRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{id}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/token/{token}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/token/{token}/start": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/sessions/{sessionId}/complete": {
      "post": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{interviewId}/sessions": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewSessionResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewSessionResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewSessionResponseDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/sessions/{sessionId}": {
      "get": {
        "tags": [
          "Interviews"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionDetailsDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddProblemToInterviewRequestDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateInterviewRequestDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "positionName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "InterviewProblemDto": {
        "type": "object",
        "properties": {
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "type": "string",
            "nullable": true
          },
          "topic": {
            "type": "string",
            "nullable": true
          },
          "constraintsText": {
            "type": "string",
            "nullable": true
          },
          "exampleInput": {
            "type": "string",
            "nullable": true
          },
          "exampleOutput": {
            "type": "string",
            "nullable": true
          },
          "executionMode": {
            "type": "string",
            "nullable": true
          },
          "signature": {
            "$ref": "#/components/schemas/ProblemSignatureDefinitionDto"
          },
          "csharpStarterCode": {
            "type": "string",
            "nullable": true
          },
          "pythonStarterCode": {
            "type": "string",
            "nullable": true
          },
          "cppStarterCode": {
            "type": "string",
            "nullable": true
          },
          "visibleTestCases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewProblemVisibleTestCaseDto"
            },
            "nullable": true
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "InterviewProblemVisibleTestCaseDto": {
        "type": "object",
        "properties": {
          "input": {
            "type": "string",
            "nullable": true
          },
          "expectedOutput": {
            "type": "string",
            "nullable": true
          },
          "orderIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "InterviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "positionName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "durationMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "accessToken": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "interviewerId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "problems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewProblemDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InterviewSessionDetailsDto": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/InterviewSessionResponseDto"
          },
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionResponseDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InterviewSessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "totalScore": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProblemSignatureDefinitionDto": {
        "type": "object",
        "properties": {
          "inputBindingMode": {
            "type": "string",
            "nullable": true
          },
          "csharpMethodName": {
            "type": "string",
            "nullable": true
          },
          "pythonMethodName": {
            "type": "string",
            "nullable": true
          },
          "cppMethodName": {
            "type": "string",
            "nullable": true
          },
          "returnType": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemSignatureParameterDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProblemSignatureParameterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubmissionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "problemId": {
            "type": "string",
            "format": "uuid"
          },
          "interviewSessionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "sourceCode": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "passedTests": {
            "type": "integer",
            "format": "int32"
          },
          "totalTests": {
            "type": "integer",
            "format": "int32"
          },
          "executionTimeMs": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "memoryKb": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "executionOutput": {
            "type": "string",
            "nullable": true
          },
          "aiFeedback": {
            "type": "string",
            "nullable": true
          },
          "aiFeedbackSource": {
            "type": "string",
            "nullable": true
          },
          "aiFeedbackStatus": {
            "type": "string",
            "nullable": true
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}